+++ src/AutoRest.CSharp/Common/AutoRest/Plugins/CSharpGen.cs
@@ -86,7 +86,7 @@ namespace AutoRest.CSharp.AutoRest.Plugins
             }
             if (Configuration.SaveInputs)
             {
-                await autoRest.WriteFile("Configuration.json", StandaloneGeneratorRunner.SaveConfiguration(), "source-file-csharp");
+                await autoRest.WriteFile("Configuration.json", Configuration.SaveConfiguration(), "source-file-csharp");
                 await autoRest.WriteFile("CodeModel.yaml", codeModelYaml, "source-file-csharp");
             }
 
@@ -110,7 +110,7 @@ namespace AutoRest.CSharp.AutoRest.Plugins
                     if (Configuration.SaveInputs)
                     {
                         // We are unsuspectingly crashing, so output anything that might help us reproduce the issue
-                        File.WriteAllText(Path.Combine(Configuration.OutputFolder, "Configuration.json"), StandaloneGeneratorRunner.SaveConfiguration());
+                        File.WriteAllText(Path.Combine(Configuration.OutputFolder, "Configuration.json"), Configuration.SaveConfiguration());
                         File.WriteAllText(Path.Combine(Configuration.OutputFolder, "CodeModel.yaml"), codeModelYaml);
                     }
                 }
